Multi-format Wiki Support
This is to help the script fit more geo templates.

To solve the inconsistency of coordinate system, parse the coordinate from the URL if possible. If the template doesn't use the geohack link, we try to get the internal built map link. The internal map is built with script, so add a mouse move event to trigger the calculation.